% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunator-package.R
\docType{package}
\name{delaunator-package}
\alias{delaunator}
\alias{delaunator-package}
\title{delaunator: R Wrapper of C++ 'delaunator'}
\description{
Provides fast Delaunay triangulation of xy point sets, using the 'delaunator' header library available at \url{https://github.com/abellgithub/delaunator-cpp}.

Delaunator
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abellgithub/delaunator-cpp}
  \item \url{https://github.com/mdsumner/delaunator1}
  \item Report bugs at \url{https://github.com/mdsumner/delaunator1/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Vladi Bilonenko [copyright holder]
}

}
\keyword{internal}
