% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunator_triangulate.R
\name{delaunator_triangulate}
\alias{delaunator_triangulate}
\title{Delaunay triangulate with 'delaunator'}
\usage{
delaunator_triangulate(x, y = NULL)
}
\arguments{
\item{x}{x coord (or xy-coords)}

\item{y}{y coord (or NULL, default)}
}
\value{
matrix of triplet indices, one triangle per row
}
\description{
Straightforward point-based triangulation.
}
\examples{
idx <- delaunator_triangulate(xy <- cbind(1:100, rnorm(100)))
plot(xy, pch = "+", cex = 2, col = "firebrick")
polygon(xy[t(cbind(idx, NA)), ])
}
